+++ 3rd_semester/hw03_tanks/01_Tanks/gun.h
@@ -7,14 +7,14 @@
 class Gun : public QGraphicsItem
 {
 public:
-	Gun(qreal width, qreal height, QColor color);
+	Gun(int width, int height, QColor color);
 	QRectF boundingRect() const;
 	void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget);
 
 
 private:
-	qreal width;
-	qreal height;
+	int width;
+	int height;
 	QColor color;
 };
