+++ 3rd_semester/hw03_tanks/01_Tanks/missile.h
@@ -17,7 +17,7 @@ public:
 	const QString &getName();
 
 protected:
-	qreal radius;
+	int radius;
 	int weight;
 	QString name;
 	Explosion *explosion;
